Use JSON API plugin instead of bundling outdated JSON library
https://plugins.jenkins.io/json-api/dependencies/ describes the technique.

https://docs.google.com/document/d/1PKYIpPlRVGsBqrz0Ob1Cv3cefOZ5j2xtGZdWs27kLuw/edit#heading=h.bjcwbz77lxkb
gives more details.

Older versions of the JSON library (like this one) used a unique
licensing scheme that has been changed in newer releases.  The JSON
API plugin uses a newer release.  Using the JSON API plugin removes the
version management chore from the plugin maintainer and leaves it with
the maintainers of the plugin bill of materials.